requirements.txt
numpy
opencv-python
Pillow
tqdm
click
requests
beautifulsoup4
easyocr
python-Levenshtein
openai
langchain-core
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
torch
torchvision
transformers
clip-retrieval
googletrans==4.0.0-rc1
gradio
fastapi
uvicorn
pydantic
wandb
gdown
ftfy
regex
llava